C++ Matrix class
C++ Matrix class ready for wrappers to read over. Still not 100% finished:
Errors with print() and transpose() functions. (They are unnecessary so we might just delete them).
Have not coded CreateIdentity() yet, this will come in a later commit.
Check out the README.txt for more details. This can also be embellished further.